requirements.txt
cloudpickle
chex
dm-haiku
dm-pix
jax
jaxlib
numpy
optax
git+https://github.com/lkhromykh/rltools.git@dev
git+https://github.com/RQC-Robotics/ur5-env.git@dev
